chore: convert fileStorage utility to TypeScript
